Extract throttle functionality from HandleContext
