+++ hedera-node/hedera-app-spi/src/main/java/com/hedera/node/app/spi/throttle/ThrottleAdviser.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hedera.node.app.spi.throttle;
+
+import com.hedera.hapi.node.base.HederaFunctionality;
+
+public interface ThrottleAdviser {
+
+    /**
+     * Verifies if the throttle in this operation context has enough capacity to handle the given number of the
+     * given function at the given time. (The time matters because we want to consider how much
+     * will have leaked between now and that time.)
+     *
+     * @param n the number of the given function
+     * @param function the function
+     * @return true if the system should throttle the given number of the given function
+     * at the instant for which throttling should be calculated
+     */
+    boolean shouldThrottleNOfUnscaled(int n, HederaFunctionality function);
+
+    /**
+     * For each following child transaction consumes the capacity
+     * required for that child transaction in the consensus throttle buckets.
+     *
+     * @return true if all the child transactions were allowed through the throttle consideration, false otherwise.
+     */
+    boolean hasThrottleCapacityForChildTransactions();
+}
