+++ hedera-node/hedera-app-spi/src/main/java/com/hedera/node/app/spi/workflows/HandleContext.java
@@ -32,6 +32,7 @@ import com.hedera.node.app.spi.records.RecordCache;
 import com.hedera.node.app.spi.signatures.SignatureVerification;
 import com.hedera.node.app.spi.signatures.VerificationAssistant;
 import com.hedera.node.app.spi.store.StoreFactory;
+import com.hedera.node.app.spi.throttle.ThrottleAdviser;
 import com.hedera.node.app.spi.validation.AttributeValidator;
 import com.hedera.node.app.spi.validation.ExpiryValidator;
 import com.hedera.node.app.spi.workflows.record.ExternalizedRecordCustomizer;
@@ -548,24 +549,12 @@ public interface HandleContext {
     void revertRecordsFrom(@NonNull RecordListCheckPoint recordListCheckPoint);
 
     /**
-     * Verifies if the throttle in this operation context has enough capacity to handle the given number of the
-     * given function at the given time. (The time matters because we want to consider how much
-     * will have leaked between now and that time.)
+     * Returns the {@link ThrottleAdviser} for this transaction, which provides information about throttles.
      *
-     * @param n the number of the given function
-     * @param function the function
-     * @return true if the system should throttle the given number of the given function
-     * at the instant for which throttling should be calculated
+     * @return the {@link ThrottleAdviser} for this transaction
      */
-    boolean shouldThrottleNOfUnscaled(int n, HederaFunctionality function);
-
-    /**
-     * For each following child transaction consumes the capacity
-     * required for that child transaction in the consensus throttle buckets.
-     *
-     * @return true if all the child transactions were allowed through the throttle consideration, false otherwise.
-     */
-    boolean hasThrottleCapacityForChildTransactions();
+    @NonNull
+    ThrottleAdviser throttleAdviser();
 
     /**
      * Create a checkpoint for the current childRecords.
