+++ hedera-node/hedera-app-spi/src/main/java/module-info.java
@@ -16,6 +16,7 @@ module com.hedera.node.app.spi {
     exports com.hedera.node.app.spi.records;
     exports com.hedera.node.app.spi.signatures;
     exports com.hedera.node.app.spi.store;
+    exports com.hedera.node.app.spi.throttle;
     exports com.hedera.node.app.spi.validation;
     exports com.hedera.node.app.spi.workflows.record;
     exports com.hedera.node.app.spi.authorization;
