+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/workflows/handle/flow/DispatchHandleContext.java
@@ -16,16 +16,12 @@
 
 package com.hedera.node.app.workflows.handle.flow;
 
-import static com.hedera.hapi.node.base.HederaFunctionality.CONTRACT_CALL;
-import static com.hedera.hapi.node.base.HederaFunctionality.CONTRACT_CREATE;
-import static com.hedera.hapi.node.base.ResponseCodeEnum.SUCCESS;
 import static com.hedera.hapi.util.HapiUtils.functionOf;
 import static com.hedera.node.app.spi.workflows.HandleContext.TransactionCategory.CHILD;
 import static java.util.Collections.emptyMap;
 import static java.util.Objects.requireNonNull;
 
 import com.hedera.hapi.node.base.AccountID;
-import com.hedera.hapi.node.base.HederaFunctionality;
 import com.hedera.hapi.node.base.Key;
 import com.hedera.hapi.node.base.ResponseCodeEnum;
 import com.hedera.hapi.node.base.SignatureMap;
@@ -37,7 +33,6 @@ import com.hedera.hapi.util.UnknownHederaFunctionality;
 import com.hedera.node.app.fees.ChildFeeContextImpl;
 import com.hedera.node.app.fees.ExchangeRateManager;
 import com.hedera.node.app.fees.FeeManager;
-import com.hedera.node.app.hapi.utils.throttles.DeterministicThrottle;
 import com.hedera.node.app.ids.WritableEntityIdStore;
 import com.hedera.node.app.records.BlockRecordManager;
 import com.hedera.node.app.signature.KeyVerifier;
@@ -54,6 +49,7 @@ import com.hedera.node.app.spi.records.RecordCache;
 import com.hedera.node.app.spi.signatures.SignatureVerification;
 import com.hedera.node.app.spi.signatures.VerificationAssistant;
 import com.hedera.node.app.spi.store.StoreFactory;
+import com.hedera.node.app.spi.throttle.ThrottleAdviser;
 import com.hedera.node.app.spi.validation.AttributeValidator;
 import com.hedera.node.app.spi.validation.ExpiryValidator;
 import com.hedera.node.app.spi.workflows.ComputeDispatchFeesAsTopLevel;
@@ -65,7 +61,6 @@ import com.hedera.node.app.spi.workflows.record.ExternalizedRecordCustomizer;
 import com.hedera.node.app.spi.workflows.record.RecordListCheckPoint;
 import com.hedera.node.app.spi.workflows.record.SingleTransactionRecordBuilder;
 import com.hedera.node.app.store.StoreFactoryImpl;
-import com.hedera.node.app.throttle.NetworkUtilizationManager;
 import com.hedera.node.app.workflows.TransactionInfo;
 import com.hedera.node.app.workflows.dispatcher.TransactionDispatcher;
 import com.hedera.node.app.workflows.handle.flow.dispatch.Dispatch;
@@ -85,9 +80,7 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.time.Instant;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.function.Predicate;
 import javax.inject.Inject;
 import javax.inject.Provider;
@@ -121,7 +114,7 @@ public class DispatchHandleContext implements HandleContext, FeeContext {
     private final ChildDispatchFactory childDispatchFactory;
     private final Dispatch currentDispatch;
     private final DispatchProcessor dispatchProcessor;
-    private final NetworkUtilizationManager networkUtilizationManager;
+    private final ThrottleAdviser throttleAdviser;
     private Map<AccountID, Long> dispatchPaidRewards;
 
     @Inject
@@ -148,7 +141,7 @@ public class DispatchHandleContext implements HandleContext, FeeContext {
             @NonNull final ChildDispatchFactory childDispatchLogic,
             @NonNull final Dispatch parentDispatch,
             @NonNull final DispatchProcessor dispatchProcessor,
-            @NonNull final NetworkUtilizationManager networkUtilizationManager) {
+            @NonNull final ThrottleAdviser throttleAdviser) {
         this.consensusNow = requireNonNull(consensusNow);
         this.txnInfo = requireNonNull(transactionInfo);
         this.configuration = requireNonNull(configuration);
@@ -167,7 +160,7 @@ public class DispatchHandleContext implements HandleContext, FeeContext {
         this.childDispatchFactory = requireNonNull(childDispatchLogic);
         this.currentDispatch = requireNonNull(parentDispatch);
         this.dispatchProcessor = requireNonNull(dispatchProcessor);
-        this.networkUtilizationManager = requireNonNull(networkUtilizationManager);
+        this.throttleAdviser = requireNonNull(throttleAdviser);
         this.attributeValidator = new AttributeValidatorImpl(this);
         this.expiryValidator = new ExpiryValidatorImpl(this);
         this.dispatcher = requireNonNull(dispatcher);
@@ -512,54 +505,10 @@ public class DispatchHandleContext implements HandleContext, FeeContext {
         currentDispatch.recordListBuilder().revertChildrenFrom(recordListCheckPoint);
     }
 
+    @NonNull
     @Override
-    public boolean shouldThrottleNOfUnscaled(final int n, @NonNull final HederaFunctionality function) {
-        requireNonNull(function);
-        return networkUtilizationManager.shouldThrottleNOfUnscaled(n, function, consensusNow);
-    }
-
-    @Override
-    public boolean hasThrottleCapacityForChildTransactions() {
-        var isAllowed = true;
-        final var childRecords = currentDispatch.recordListBuilder().childRecordBuilders();
-        @Nullable List<DeterministicThrottle.UsageSnapshot> snapshotsIfNeeded = null;
-
-        for (int i = 0, n = childRecords.size(); i < n && isAllowed; i++) {
-            final var childRecord = childRecords.get(i);
-            if (Objects.equals(childRecord.status(), SUCCESS)) {
-                final var childTx = childRecord.transaction();
-                final var childTxBody = childRecord.transactionBody();
-                HederaFunctionality childTxFunctionality;
-                try {
-                    childTxFunctionality = functionOf(childTxBody);
-                } catch (UnknownHederaFunctionality e) {
-                    throw new IllegalStateException("Invalid transaction body " + childTxBody, e);
-                }
-
-                if (childTxFunctionality == CONTRACT_CREATE || childTxFunctionality == CONTRACT_CALL) {
-                    continue;
-                }
-                if (snapshotsIfNeeded == null) {
-                    snapshotsIfNeeded = networkUtilizationManager.getUsageSnapshots();
-                }
-
-                final var childTxInfo = TransactionInfo.from(
-                        childTx,
-                        childTxBody,
-                        childTx.sigMapOrElse(SignatureMap.DEFAULT),
-                        childTx.signedTransactionBytes(),
-                        childTxFunctionality);
-                final var shouldThrottleTxn = networkUtilizationManager.shouldThrottle(
-                        childTxInfo, currentDispatch.stack().peek(), consensusNow);
-                if (shouldThrottleTxn) {
-                    isAllowed = false;
-                }
-            }
-        }
-        if (!isAllowed) {
-            networkUtilizationManager.resetUsageThrottlesTo(snapshotsIfNeeded);
-        }
-        return isAllowed;
+    public ThrottleAdviser throttleAdviser() {
+        return throttleAdviser;
     }
 
     @NonNull
