+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/workflows/handle/flow/dispatch/child/modules/ChildDispatchModule.java
@@ -42,12 +42,12 @@ import com.hedera.node.app.spi.fees.ResourcePriceCalculator;
 import com.hedera.node.app.spi.metrics.StoreMetricsService;
 import com.hedera.node.app.spi.records.RecordCache;
 import com.hedera.node.app.spi.store.StoreFactory;
+import com.hedera.node.app.spi.throttle.ThrottleAdviser;
 import com.hedera.node.app.spi.workflows.HandleContext;
 import com.hedera.node.app.store.ReadableStoreFactory;
 import com.hedera.node.app.store.ServiceApiFactory;
 import com.hedera.node.app.store.StoreFactoryImpl;
 import com.hedera.node.app.store.WritableStoreFactory;
-import com.hedera.node.app.throttle.NetworkUtilizationManager;
 import com.hedera.node.app.workflows.TransactionInfo;
 import com.hedera.node.app.workflows.dispatcher.TransactionDispatcher;
 import com.hedera.node.app.workflows.handle.TriggeredFinalizeContext;
@@ -217,15 +217,13 @@ public interface ChildDispatchModule {
             @NonNull final WritableEntityIdStore entityIdStore,
             @NonNull final TransactionDispatcher dispatcher,
             @NonNull final RecordCache recordCache,
-            @NonNull final WritableStoreFactory writableStoreFactory,
-            @NonNull final ServiceApiFactory serviceApiFactory,
             @NonNull final NetworkInfo networkInfo,
             @NonNull final SingleTransactionRecordBuilderImpl recordBuilder,
             @NonNull final Provider<ChildDispatchComponent.Factory> childDispatchFactory,
             @NonNull final ChildDispatchFactory childDispatchLogic,
             @NonNull final ChildDispatchComponent dispatch,
             @NonNull final DispatchProcessor dispatchProcessor,
-            @NonNull final NetworkUtilizationManager networkUtilizationManager) {
+            @NonNull final ThrottleAdviser throttleAdviser) {
         return new DispatchHandleContext(
                 recordBuilder.consensusNow(),
                 transactionInfo,
@@ -249,6 +247,6 @@ public interface ChildDispatchModule {
                 childDispatchLogic,
                 dispatch,
                 dispatchProcessor,
-                networkUtilizationManager);
+                throttleAdviser);
     }
 }
