+++ hedera-node/hedera-app/src/test/java/com/hedera/node/app/workflows/handle/flow/DispatchHandleContextTest.java
@@ -16,12 +16,9 @@
 
 package com.hedera.node.app.workflows.handle.flow;
 
-import static com.hedera.hapi.node.base.HederaFunctionality.CONTRACT_CALL;
 import static com.hedera.hapi.node.base.HederaFunctionality.CRYPTO_TRANSFER;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.INSUFFICIENT_ACCOUNT_BALANCE;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.INVALID_TRANSACTION_BODY;
-import static com.hedera.hapi.node.base.ResponseCodeEnum.REVERTED_SUCCESS;
-import static com.hedera.hapi.node.base.ResponseCodeEnum.SUCCESS;
 import static com.hedera.hapi.node.base.ResponseCodeEnum.UNRESOLVABLE_REQUIRED_SIGNERS;
 import static com.hedera.hapi.node.base.SubType.TOKEN_NON_FUNGIBLE_UNIQUE_WITH_CUSTOM_FEES;
 import static com.hedera.hapi.util.HapiUtils.functionOf;
@@ -63,7 +60,6 @@ import com.hedera.hapi.node.base.TokenTransferList;
 import com.hedera.hapi.node.base.Transaction;
 import com.hedera.hapi.node.base.TransactionID;
 import com.hedera.hapi.node.consensus.ConsensusSubmitMessageTransactionBody;
-import com.hedera.hapi.node.contract.ContractCallTransactionBody;
 import com.hedera.hapi.node.state.common.EntityNumber;
 import com.hedera.hapi.node.state.token.Account;
 import com.hedera.hapi.node.token.CryptoTransferTransactionBody;
@@ -76,7 +72,6 @@ import com.hedera.node.app.ids.EntityIdService;
 import com.hedera.node.app.ids.WritableEntityIdStore;
 import com.hedera.node.app.records.BlockRecordManager;
 import com.hedera.node.app.service.token.TokenService;
-import com.hedera.node.app.service.token.impl.WritableAccountStore;
 import com.hedera.node.app.service.token.records.CryptoCreateRecordBuilder;
 import com.hedera.node.app.services.ServiceScopeLookup;
 import com.hedera.node.app.signature.KeyVerifier;
@@ -92,6 +87,7 @@ import com.hedera.node.app.spi.fixtures.state.MapWritableStates;
 import com.hedera.node.app.spi.metrics.StoreMetricsService;
 import com.hedera.node.app.spi.signatures.SignatureVerification;
 import com.hedera.node.app.spi.signatures.VerificationAssistant;
+import com.hedera.node.app.spi.throttle.ThrottleAdviser;
 import com.hedera.node.app.spi.workflows.ComputeDispatchFeesAsTopLevel;
 import com.hedera.node.app.spi.workflows.HandleContext;
 import com.hedera.node.app.spi.workflows.HandleException;
@@ -105,7 +101,6 @@ import com.hedera.node.app.store.ReadableStoreFactory;
 import com.hedera.node.app.store.ServiceApiFactory;
 import com.hedera.node.app.store.StoreFactoryImpl;
 import com.hedera.node.app.store.WritableStoreFactory;
-import com.hedera.node.app.throttle.NetworkUtilizationManager;
 import com.hedera.node.app.workflows.TransactionInfo;
 import com.hedera.node.app.workflows.dispatcher.TransactionDispatcher;
 import com.hedera.node.app.workflows.handle.flow.dispatch.Dispatch;
@@ -156,19 +151,12 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 public class DispatchHandleContextTest extends StateTestBase implements Scenarios {
-    private static final long GAS_LIMIT = 456L;
     private static final Fees FEES = new Fees(1L, 2L, 3L);
     public static final Instant CONSENSUS_NOW = Instant.ofEpochSecond(1_234_567L, 890);
     private static final AccountID PAYER_ACCOUNT_ID =
             AccountID.newBuilder().accountNum(1_234).build();
     private static final AccountID NODE_ACCOUNT_ID =
             AccountID.newBuilder().accountNum(3).build();
-    private static final TransactionBody CONTRACT_CALL_TXN_BODY = TransactionBody.newBuilder()
-            .transactionID(
-                    TransactionID.newBuilder().accountID(PAYER_ACCOUNT_ID).build())
-            .contractCall(
-                    ContractCallTransactionBody.newBuilder().gas(GAS_LIMIT).build())
-            .build();
     private static final SignatureVerification FAILED_VERIFICATION =
             new SignatureVerificationImpl(Key.DEFAULT, Bytes.EMPTY, false);
     private static final TransactionBody MISSING_FUNCTION_TXN_BODY = TransactionBody.newBuilder()
@@ -182,8 +170,6 @@ public class DispatchHandleContextTest extends StateTestBase implements Scenario
             .build();
     private static final TransactionInfo CRYPTO_TRANSFER_TXN_INFO = new TransactionInfo(
             Transaction.DEFAULT, CRYPTO_TRANSFER_TXN_BODY, SignatureMap.DEFAULT, Bytes.EMPTY, CRYPTO_TRANSFER);
-    private static final TransactionInfo CONTRACT_CALL_TXN_INFO = new TransactionInfo(
-            Transaction.DEFAULT, CONTRACT_CALL_TXN_BODY, SignatureMap.DEFAULT, Bytes.EMPTY, CONTRACT_CALL);
 
     @Mock
     private KeyVerifier verifier;
@@ -219,19 +205,13 @@ public class DispatchHandleContextTest extends StateTestBase implements Scenario
     private SignatureVerification verification;
 
     @Mock
-    private NetworkUtilizationManager networkUtilizationManager;
+    private ThrottleAdviser throttleAdviser;
 
     @Mock
     private StoreMetricsService storeMetricsService;
 
     @Mock
     private WritableEntityIdStore entityIdStore;
-
-    @Mock
-    private SingleTransactionRecordBuilderImpl oneChildBuilder;
-
-    @Mock
-    private SingleTransactionRecordBuilderImpl twoChildBuilder;
 
     @Mock
     private Provider<ChildDispatchComponent.Factory> childDispatchProvider;
@@ -267,13 +247,7 @@ public class DispatchHandleContextTest extends StateTestBase implements Scenario
     private SavepointStackImpl stack;
 
     @Mock
-    private WrappedHederaState wrappedHederaState;
-
-    @Mock
     private WritableStoreFactory writableStoreFactory;
-
-    @Mock
-    private WritableAccountStore writableAccountStore;
 
     @Mock
     private VerificationAssistant assistant;
@@ -410,7 +384,7 @@ public class DispatchHandleContextTest extends StateTestBase implements Scenario
             childDispatchFactory,
             parentDispatch,
             dispatchProcessor,
-            networkUtilizationManager
+            throttleAdviser
         };
 
         final var constructor = DispatchHandleContext.class.getConstructors()[0];
@@ -926,43 +900,6 @@ public class DispatchHandleContextTest extends StateTestBase implements Scenario
         verify(recordListBuilder).revertChildrenFrom(checkpoint);
     }
 
-    @Test
-    void dispatchesThrottlingN() {
-        subject.shouldThrottleNOfUnscaled(2, CRYPTO_TRANSFER);
-        verify(networkUtilizationManager).shouldThrottleNOfUnscaled(2, CRYPTO_TRANSFER, CONSENSUS_NOW);
-    }
-
-    @Test
-    void allowsThrottleCapacityForChildrenIfNoneShouldThrottle() {
-        given(parentDispatch.recordListBuilder()).willReturn(recordListBuilder);
-        given(recordListBuilder.childRecordBuilders()).willReturn(List.of(oneChildBuilder, twoChildBuilder));
-        given(oneChildBuilder.status()).willReturn(SUCCESS);
-        given(oneChildBuilder.transaction()).willReturn(CRYPTO_TRANSFER_TXN_INFO.transaction());
-        given(oneChildBuilder.transactionBody()).willReturn(CRYPTO_TRANSFER_TXN_INFO.txBody());
-        given(twoChildBuilder.status()).willReturn(REVERTED_SUCCESS);
-        given(parentDispatch.stack()).willReturn(stack);
-        given(stack.peek()).willReturn(wrappedHederaState);
-
-        assertThat(subject.hasThrottleCapacityForChildTransactions()).isTrue();
-    }
-
-    @Test
-    void doesntAllowThrottleCapacityForChildrenIfOneShouldThrottle() {
-        given(parentDispatch.recordListBuilder()).willReturn(recordListBuilder);
-        given(recordListBuilder.childRecordBuilders()).willReturn(List.of(oneChildBuilder, twoChildBuilder));
-        given(oneChildBuilder.status()).willReturn(SUCCESS);
-        given(oneChildBuilder.transaction()).willReturn(CONTRACT_CALL_TXN_INFO.transaction());
-        given(oneChildBuilder.transactionBody()).willReturn(CONTRACT_CALL_TXN_INFO.txBody());
-        given(twoChildBuilder.status()).willReturn(SUCCESS);
-        given(twoChildBuilder.transaction()).willReturn(CRYPTO_TRANSFER_TXN_INFO.transaction());
-        given(twoChildBuilder.transactionBody()).willReturn(CRYPTO_TRANSFER_TXN_INFO.txBody());
-        given(networkUtilizationManager.shouldThrottle(any(), eq(wrappedHederaState), eq(CONSENSUS_NOW)))
-                .willReturn(true);
-        given(parentDispatch.stack()).willReturn(stack);
-        given(stack.peek()).willReturn(wrappedHederaState);
-        assertThat(subject.hasThrottleCapacityForChildTransactions()).isFalse();
-    }
-
     private DispatchHandleContext createContext(final TransactionBody txBody) {
         return createContext(txBody, HandleContext.TransactionCategory.USER);
     }
@@ -1002,7 +939,7 @@ public class DispatchHandleContextTest extends StateTestBase implements Scenario
                 childDispatchFactory,
                 parentDispatch,
                 dispatchProcessor,
-                networkUtilizationManager);
+                throttleAdviser);
     }
 
     private void mockNeeded() {
