+++ hedera-node/hedera-smart-contract-service-impl/src/main/java/com/hedera/node/app/service/contract/impl/state/RootProxyWorldUpdater.java
@@ -106,11 +106,11 @@ public class RootProxyWorldUpdater extends ProxyWorldUpdater {
 
         if (contractsConfig.enforceCreationThrottle()) {
             final var creationCapacityIsAvailable =
-                    !context.shouldThrottleNOfUnscaled(createdContractIds.size(), CRYPTO_CREATE);
+                    !context.throttleAdviser().shouldThrottleNOfUnscaled(createdContractIds.size(), CRYPTO_CREATE);
             validateResource(creationCapacityIsAvailable, CONSENSUS_GAS_EXHAUSTED);
         }
 
-        final var childThrottleIsAvailable = context.hasThrottleCapacityForChildTransactions();
+        final var childThrottleIsAvailable = context.throttleAdviser().hasThrottleCapacityForChildTransactions();
         validateResource(childThrottleIsAvailable, CONSENSUS_GAS_EXHAUSTED);
 
         // If nonces externalization is enabled, we need to capture the updated nonces
